merge Dev to main (#5)
* feat(seo): implement technical SEO improvements

This commit implements several technical SEO improvements based on the SEO plan.

- **robots.txt:** Added a `robots.txt` file to allow search engines to crawl the site and to point them to the sitemap.
- **Page Speed:** Minified `main.css` and `main.js` to improve page load times.

* mv some members to alumni

* fix(seo): correct plugin configuration for sitemap generation

This commit corrects the plugin configuration in `_config.yml` to ensure the `jekyll-sitemap` plugin is loaded correctly. This should fix the issue with the sitemap not being generated.

This commit also includes the changes from the previous commit that was not pushed:
- feat(people): move alumni to alumni page
- feat(people): update alumni information

* deleted some alumni from current people